pyproject.toml
[project]
name = "agent-loop"
version = "0.1.0"
description = "Enterprise-grade Agent-Loop service with pluggable skills"
requires-python = ">=3.10"
dependencies = [
    # Web framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",

    # LangGraph ecosystem
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",

    # MCP
    "mcp>=0.1.0",

    # Data validation and configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",

    # Database
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "redis>=5.0.0",

    # Logging
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "anyio>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "black>=23.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]